move http_worker to http submodule
