+++ distributed/http/__init__.py
@@ -1,0 +1,1 @@
+from .worker import HTTPWorker
