+++ distributed/http/core.py
@@ -1,23 +1,14 @@
+
 import json
 
 from toolz import first
 from tornado import web
 from tornado.httpserver import HTTPServer
 
-
 class JSON(web.RequestHandler):
     def write(self, obj):
         return super(JSON, self).write(json.dumps(obj))
 
-class Info(JSON):
-    def initialize(self, worker):
-        self.worker = worker
-
-    def get(self):
-        resp = {'ncores': self.worker.ncores,
-                'nkeys': len(self.worker.data),
-                'status': self.worker.status}
-        self.write(resp)
 
 class MyApp(HTTPServer):
     @property
@@ -40,8 +31,3 @@ class MyApp(HTTPServer):
                 else:
                     logger.info('Randomly assigned port taken for %s. Retrying',
                                 type(self).__name__)
-
-def HTTPWorker(worker):
-    application = MyApp(web.Application([
-        (r'/info.json', Info, {'worker': worker})]))
-    return application
